- Fix ConditionModel & LanguageModel in for sale - Make LowestAskcost to double

+++ app/src/main/java/com/gthr/gthrcollect/model/network/firebaserealtimedb/FunkoModel.kt
@@ -39,7 +39,7 @@ data class FunkoModel(
     var license: String? = "",
     @get:PropertyName(FirebaseRealtimeDatabase.LOWEST_ASK_COST)
     @set:PropertyName(FirebaseRealtimeDatabase.LOWEST_ASK_COST)
-    var lowestAskCost: Int? = -1,
+    var lowestAskCost: Double? = -1.0,
     @get:PropertyName(FirebaseRealtimeDatabase.LOWEST_ASK_ID)
     @set:PropertyName(FirebaseRealtimeDatabase.LOWEST_ASK_ID)
     var lowestAskID: String? = "",
